Implement stroke fill
We implement our own fill algorithm to make it easier to handle the
resulting shapes with ImageData access / without the predefined
canvas functionality.

We implement a simple old-school pixel circle algorithm that spits out
horizontal spans and then write a merge algorithm to merge these
multiple spans if necessary to a minimal number of spans for that
shape.

Adds tests for the stroke merging